Payment info update and Activities section update
In the activities section when clicking a checkbox or unchecking the final price updates accordingly.  If a selection has a competing time slot, then the other selection becomes disabled so a user cannot sign up for it.  In the payment section when a selection is made, then the other two sections are hidden.